jf: Keep SIM state on LPM mode
When the device boots up in airplane mode, the card state is never
queried and the device thinks there is no SIM.

To fix this, instead of syncing the card state only on radio on,
sync the card state on radio available.

It is still not confirmed whether this fixes the issue as the issue
is pretty erratic and happens only sometimes, but the issue doesn't
seem to happen anymore with this change.

This also matches the prop in both GPE GUEUDPF1 and TW XXUHPF4.

Change-Id: Ifade7df9b42802b2a57f47291998876643d6a8fc
Signed-off-by: Paul Keith <javelinanddart@gmail.com>